HR&Finances More insert, changed data types

+++ RoofingCompany.sql
@@ -360,8 +360,8 @@ PESEL bigint,
 create table Position(
 IdPosition int not null Primary key identity(1,1),
 Workplace nvarchar(50) null,
-ValidityOfOshTraining tinyint null,
-VailidityOfMedicalExam tinyint null
+ValidityOfOshTraining int null,
+VailidityOfMedicalExam int null
 );
 
 create table Contract(
@@ -412,7 +412,7 @@ Create table Staff(
 IdStaff int primary key identity(1,1) not null,
 IdDeparment int foreign key references Department(IdDepartment),
 IdPosition int foreign key references Position(IdPosition),
-Number tinyint,
+Number int,
 DateFrom date
 );
